Remove separate shortcuts
The plugin used to have 2 shortcuts: one to auto-select current
paragraph and reflow it and the other to reflow selected text. But this
can be decided based on the presence of a selection: if a text is
selected, format the selection, otherwise format the current paragraph.

If there wasn't a selection and "format selection" was used, the current
line was selected and formatted, which isn't very intuitive and I never
used it.